requirements.txt
qrcode
Pillow

